Added new 64 bit fields nSize and nUndoSize to CBlockFileInfo and renamed previous 32 bit nSize and nUndoSize to nSizeLegacy and nUndoSizeLegacy. Also updated usages of this two fields to 64 bit.

+++ src/test/blockfile_reading_tests.cpp
@@ -29,7 +29,7 @@ namespace
         CBlockIndex& index,
         CBlockFileInfoStore& blockFileInfoStore)
     {
-        unsigned int nBlockSizeWithHeader =
+        uint64_t nBlockSizeWithHeader =
             ::GetSerializeSize(block, SER_DISK, CLIENT_VERSION)
             + BLOCKFILE_BLOCK_HEADER_SIZE;
         CDiskBlockPos blockPos;
